Catch exceptions from mirror.record_finished_package()
I am mirroring to a Samba share. For reasons unknown, writes sometimes
transiently fail with EBADF. When this happens, the Worker in question pops its
clogs, is removed from the pool and is not replaced. So after 3 (by default)
such transient errors, there are no more workers and the sync stops.

While obviously it would be best if the Samba share behaved itself, let's treat
this error the same way as any other error during syncing any particular
package and allow the rest of the sync to continue.