Add pep8 tox env
pytest-codecheckers doesn't work on Python 2.7.  The released version
fails for me with

 INTERNALERROR> AttributeError: 'PyCodeCheckItem' object has no attribute 'out'

There's unreleased code that might work [1] but it doesn't seem to
have been released since 2012.  Even when fixing it up manually, it
then goes ahead and runs checks over all the virtualenv files leading
to a bunch of failures in code you don't control.

Remove this plugin and move to a separate pep8 environment for code
checking (influenced by OpenStack's usual idiom for this)

[1] https://bitbucket.org/RonnyPfannschmidt/pytest-codecheckers/src/ffa881a66ef64576247bbdd2895b3b737e99469f/codecheckers/plugin.py?at=default&fileviewer=file-view-default#plugin.py-28